Ajout du rendu3 - mis a jour

+++ rendu3/requete.sql
@@ -131,7 +131,7 @@ GROUP BY(H.nom_gare);
 
 -- Taux de remplissage des trains
 
-select train, (nbPersonnes::float / tt.nbPlace::float) as "Taux de remplissage"
+select train, (nbPersonnes::float / tt.nbPlace) as "Taux de remplissage"
 from vNbPersParTrain nbP join train t on nbP.train = t.numero
     join TypeTrain tt on t.type = tt.nom;
